requirements.txt
Flask==2.3.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
schedule==1.2.0
secure-smtplib
cryptography==41.0.3
reportlab==4.0.8
